Add the Value enum and a basic nom parser.
The Value enum includes just the basic types of values needed to get the
interpreter to a MVP state; I chose to keep it to a minimal level for
now for ease-of-code.

The parser is *very* basic. There are no string escapes, for example, or
other things, but should also be good enough for a MVP state.